major work in progress moving generator sdk code around
