+++ lib/sdkGenHelper.js
@@ -22,7 +22,7 @@ var rimraf = require('rimraf');
 var config = require('../config');
 
 exports.performSDKGeneration = function(sdkName, sdkType, fileContent, callback) {
-
+  var self = this
   var startGenURL = config.sdkGenURL + sdkName + "/" + sdkType; // ex. ios_swift or server_swift
   request.post({
     headers: {'Accept' : 'application/json',
@@ -45,7 +45,7 @@ exports.performSDKGeneration = function(sdkName, sdkType, fileContent, callback)
     var timerId = setInterval(function () {   
     	intervalsRun += 1
     	if (intervalsRun >= 10) clearInterval(timerId);
-        getStatus()
+        getStatus();
     }, 3000) // Check status every 3 seconds
 
     function getStatus() {
@@ -74,72 +74,110 @@ exports.performSDKGeneration = function(sdkName, sdkType, fileContent, callback)
         headers: {'Accept' : 'application/zip'},
         url: config.sdkGenURL + genID
       })
-      .pipe(fs.createWriteStream(sdkName + '.zip'))
-      .on('finish', callback);
+      .on('data', function(data) {
+      	self.fs.write(self.destinationPath(sdkName + '.zip'), data);
+      })
+      .on('end', callback);
     }
   });
 }
 
 exports.extractNewContent = function(sdkName, callback) {
+	console.log("newcontent");
+	var self = this;
+	this.extract(this.destinationPath(sdkName + '.zip'), this.destinationPath('./'), function() {
+		console.log("Extracted!");
+		console.log(self.fs.exists('./' + sdkName + '/Package.swift'));
+		// var originFile = fs.readFileSync('./' + sdkName + '/Package.swift', 'utf8');
+		// var originFile = self.fs.read('./' + sdkName + '/Package.swift');
+		// console.log(originFile);
+		// var originMatches = originFile.match(/\.\bPackage\b.*/g);
+		// var packages = '';
 
-	fs.createReadStream(sdkName + '.zip')
-	  .pipe(unzip.Extract({ path: '.' })
-	  .on('close', function () {
+		// if(originMatches != null && originMatches.length > 0) {
 
-		var originFile = fs.readFileSync('./' + sdkName + '/Package.swift', 'utf8');
-		var originMatches = originFile.match(/\.\bPackage\b.*/g);
-		var packages = '';
+		// 	var index; 
+		// 	for (index = 0; index < originMatches.length; index++) {
+		// 		packages += originMatches[index];
+		// 		if(index + 1 == originMatches.length) packages += ',';
+		// 		packages += '\n\t\t\t\t'
+		// 	}
+		// } else {
+		// 	console.warn('Failed to find matches for Package dependencies.');
+		// }
 
-		if(originMatches != null && originMatches.length > 0) {
+		// callback(sdkName, packages);
+		callback(sdkName, 'apackage');
+	});
 
-			var index; 
-			for (index = 0; index < originMatches.length; index++) {
-				packages += originMatches[index];
-				if(index + 1 == originMatches.length) packages += ',';
-				packages += '\n\t\t\t\t'
-			}
-		} else {
-			console.warn('Failed to find matches for Package dependencies.');
-		}
+	// fs.createReadStream(sdkName + '.zip')
+	//   .pipe(unzip.Extract({ path: '.' })
+	//   .on('close', function () {
 
-		callback(sdkName, packages);
+	// 	var originFile = fs.readFileSync('./' + sdkName + '/Package.swift', 'utf8');
+	// 	var originMatches = originFile.match(/\.\bPackage\b.*/g);
+	// 	var packages = '';
 
-		// The following is an alternate way to get the sdk module name using Regex
-		/*var nameMatch = originFile.match(/"(.*?)"/i);
-		if(nameMatch != null && nameMatch.length >= 2) {
-			// Get group 1 result, parsing out quotes just in case
-			var desiredMatch = nameMatch[1].replace(/['"]+/g, '');
-			callback(desiredMatch, packages);
+	// 	if(originMatches != null && originMatches.length > 0) {
 
-		} else {
-			console.log("Failed to find match for Package name.")
-			callback('', packages);
-		}*/
-	}));
+	// 		var index; 
+	// 		for (index = 0; index < originMatches.length; index++) {
+	// 			packages += originMatches[index];
+	// 			if(index + 1 == originMatches.length) packages += ',';
+	// 			packages += '\n\t\t\t\t'
+	// 		}
+	// 	} else {
+	// 		console.warn('Failed to find matches for Package dependencies.');
+	// 	}
+
+	// 	callback(sdkName, packages);
+
+	// 	// The following is an alternate way to get the sdk module name using Regex
+	// 	/*var nameMatch = originFile.match(/"(.*?)"/i);
+	// 	if(nameMatch != null && nameMatch.length >= 2) {
+	// 		// Get group 1 result, parsing out quotes just in case
+	// 		var desiredMatch = nameMatch[1].replace(/['"]+/g, '');
+	// 		callback(desiredMatch, packages);
+
+	// 	} else {
+	// 		console.log("Failed to find match for Package name.")
+	// 		callback('', packages);
+	// 	}*/
+	// }));
 
 }
 
 exports.integrateServerSDK = function(sdkName, callback) {
+	// since files haven't been created yet, must use yeoman in memory file system
+	var origin = this.destinationPath(sdkName + '/Sources/');
+	console.log(origin);
+	var destination = this.destinationPath('Sources/' + sdkName);
+	console.log("path: " + destination)
+	// if (!this.fs.exists(destination)){
+	//   fs.mkdirSync(destination);
+	// }
 
-	var destination = './Sources/' + sdkName;
-	if (!fs.existsSync(destination)){
-	  fs.mkdirSync(destination);
-	}
+	this.fs.copy(origin, destination);
+	console.log(this.fs.exists(sdkName + '.zip'));
+	this.fs.delete(this.destinationPath(sdkName + '.zip'));
+	// this.fs.delete(sdkName);
+
+	callback();
 
 	// Write source files to app's Source folder
-	var rsync = new Rsync()
-	.flags('av')
-	.source('./' + sdkName + '/Sources/')
-	.destination(destination);
-	// Execute the resync command
-	rsync.execute(function(error, code, cmd) {
-		if(error) {
-		  console.error('Finished Rsync with error ' + error);
-		}
-		// Remove unecessary files/folders
-		fs.unlinkSync('./' + sdkName + '.zip');
-		rimraf('./' + sdkName, function () { 
-			callback();
-		});
-	});
+	// var rsync = new Rsync()
+	// .flags('av')
+	// .source('./' + sdkName + '/Sources/')
+	// .destination(destination);
+	// // Execute the resync command
+	// rsync.execute(function(error, code, cmd) {
+	// 	if(error) {
+	// 	  console.error('Finished Rsync with error ' + error);
+	// 	}
+	// 	// Remove unecessary files/folders
+	// 	fs.unlinkSync('./' + sdkName + '.zip');
+	// 	rimraf('./' + sdkName, function () { 
+	// 		callback();
+	// 	});
+	// });
 };
