+++ refresh/index.js
@@ -27,6 +27,10 @@ var actions = require('../lib/actions');
 
 var util = require('util');
 var swaggerize = require('./fromswagger/swaggerize');
+var sdkHelper = require('../lib/sdkGenHelper');
+var performSDKGeneration = sdkHelper.performSDKGeneration;
+var extractNewContent = sdkHelper.extractNewContent;
+var integrateServerSDK = sdkHelper.integrateServerSDK;
 
 module.exports = generators.Base.extend({
   constructor: function() {
@@ -192,6 +196,16 @@ module.exports = generators.Base.extend({
       if (this.spec.fromSwagger && typeof(this.spec.fromSwagger) === 'string') {
         this.fromSwagger = this.spec.fromSwagger;
       }
+
+      if (this.exampleEndpoints) {
+        if (this.fromSwagger) {
+          this.env.error('Only one of: swagger file and example endpoints allowed');
+        }
+        this.fromSwagger = this.templatePath('common', 'productSwagger.yaml');
+      }
+
+      // Swagger file paths for server SDKs
+      this.serverSwaggerFiles = this.spec.serverSwaggerFiles || [];
 
       // Swagger hosting
       this.hostSwagger = (this.spec.hostSwagger === true);
@@ -657,6 +671,67 @@ module.exports = generators.Base.extend({
   },
 
   writing: {
+    generateSDKs: function() {
+      if(!this.fromSwagger && !this.serverSwaggerFiles) return;
+      this.log(chalk.green('Generating SDK(s) from swagger file(s)...'));
+      var done = this.async();
+      var self = this; // local copy to be used in callbacks
+      console.log("fromSwagger: " + this.fromSwagger);
+      console.log("serverSwaggerFiles: " + this.serverSwaggerFiles.length);
+      // Cover the different cases
+      if(self.fromSwagger && self.serverSwaggerFiles.length <= 0) {
+        geniOS(done);
+      } else if(self.fromSwagger && self.serverSwaggerFiles.length > 0) {
+        geniOS(function() {
+          genServer(done);
+        })
+      } else if(!self.fromSwagger && self.serverSwaggerFiles.length > 0) {
+        genServer(done);
+      }
+
+      function geniOS(callback) {
+        swaggerize.parse.call(self, self.fromSwagger, function(loadedApi, parsed) {
+          performSDKGeneration(self.appname + '_iOS_SDK', 'ios_swift', JSON.stringify(loadedApi), function () {
+            callback();
+          });
+        });
+      }
+
+      function genServer(callback) {
+        var numFinished = 0;
+        for (var index = 0; index < self.serverSwaggerFiles.length; index++) {
+        
+          swaggerize.parse.call(self, self.serverSwaggerFiles[index], function(loadedApi, parsed) {
+
+            if (loadedApi['info']['title'] == undefined) {
+              this.env.error(chalk.red(err));
+            } else {
+              var sdkName = loadedApi['info']['title'].replace(/ /g, '_') + '_ServerSDK';
+              performSDKGeneration.call(self, sdkName, 'server_swift', JSON.stringify(loadedApi), function() {
+
+                extractNewContent.call(self, sdkName, function(sdkTargets, sdkPackages) {
+
+                  if(sdkTargets.length > 0) {
+                    if(self.sdkTargets === undefined) {
+                      self.sdkTargets = [];
+                    }
+                    self.sdkTargets.push(sdkTargets);
+                  }
+                  if(sdkPackages.length > 0) {
+                    self.sdkPackages = sdkPackages;
+                  }
+                  numFinished += 1;
+                  if(numFinished === self.serverSwaggerFiles.length) {
+                    callback();
+                  }
+                })
+              })
+            }
+          });
+        }
+      }
+    },
+
     createCommonFiles: function() {
 
       // Check if we should create generator metadata files
@@ -1190,6 +1265,25 @@ module.exports = generators.Base.extend({
           }
         )
       });
+    },
+
+    writeServerSDKConnection: function() {
+      if(this.fromSwagger) {
+        ignoreFile('/' + this.appname + '_iOS_SDK*');
+      }
+      return; // TODO: remove after testing
+      if(!this.sdkTargets) return;
+      var done = this.async();
+
+      var numFinished = 0, length = this.sdkTargets.length; 
+      for (var index = 0; index < length; index++) {
+        integrateServerSDK.call(this, this.sdkTargets[index], function() {
+          numFinished += 1;
+          if(numFinished === length) {
+            done();
+          }
+        });
+      }
     }
   },
 });
